# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Solar Installation CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service:
    get:
      operationId: listServiceCalls
      tags: [service]
      summary: List service calls
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: assignedEngineerId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Service calls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceCall"
    post:
      operationId: createServiceCall
      tags: [service]
      summary: Create a service call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCallInput"
      responses:
        "201":
          description: Created service call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceCall"
  /service/{id}:
    get:
      operationId: getServiceCall
      tags: [service]
      summary: Get service call by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceCall"
    patch:
      operationId: updateServiceCall
      tags: [service]
      summary: Update service call
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCallUpdate"
      responses:
        "200":
          description: Updated service call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceCall"
    delete:
      operationId: deleteServiceCall
      tags: [service]
      summary: Delete service call
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ACTIVITIES ──────────────────────────────────────────────────────────────
...
  /reports/leads:
    get:
      operationId: getLeadsReport
      tags: [reports]
      summary: Leads report
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsReport"
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Sales report
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /reports/finance:
    get:
      operationId: getFinanceReport
      tags: [reports]
      summary: Finance report
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Finance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceReport"
  /reports/service:
    get:
      operationId: getServiceReport
      tags: [reports]
      summary: Service report
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Service report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceReport"
...
  /settings:
    get:
      operationId: listSettings
      tags: [settings]
      summary: Get all settings
      responses:
        "200":
          description: Settings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Setting"
    post:
      operationId: bulkUpdateSettings
      tags: [settings]
      summary: Bulk update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Setting"
  /settings/{key}:
    get:
      operationId: getSetting
      tags: [settings]
      summary: Get setting by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"
    put:
      operationId: upsertSetting
      tags: [settings]
      summary: Create or update a setting
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingInput"
      responses:
        "200":
          description: Upserted setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"

  # ─── BACKUP ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [admin, sales, finance, warehouse, engineer]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, email, role, isActive, createdAt, updatedAt]
...
    ServiceCall:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: ["string", "null"]
        issueDescription:
          type: string
        status:
          type: string
        priority:
          type: string
        assignedEngineerId:
          type: ["integer", "null"]
        assignedEngineer:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        closureNotes:
          type: ["string", "null"]
        scheduledDate:
          type: ["string", "null"]
        closedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, customerName, customerPhone, issueDescription, status, priority, createdAt, updatedAt]
    ServiceCallInput:
      type: object
      properties:
        projectId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        issueDescription:
          type: string
        status:
          type: string
        priority:
          type: string
        assignedEngineerId:
          type: integer
        scheduledDate:
          type: string
      required: [customerName, customerPhone, issueDescription]
    ServiceCallUpdate:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        issueDescription:
          type: string
        status:
          type: string
        priority:
          type: string
        assignedEngineerId:
          type: ["integer", "null"]
        closureNotes:
          type: string
        scheduledDate:
          type: string
...
    LeadsReport:
      type: object
      properties:
        totalLeads:
          type: integer
        newLeads:
          type: integer
        convertedLeads:
          type: integer
        bySource:
          type: array
          items:
            type: object
            properties:
              source:
                type: string
              count:
                type: integer
            required: [source, count]
        byStage:
          type: array
          items:
            type: object
            properties:
              stage:
                type: string
              count:
                type: integer
            required: [stage, count]
        daily:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
            required: [date, count]
      required: [totalLeads, newLeads, convertedLeads, bySource, byStage, daily]
    SalesReport:
      type: object
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        bySalesPerson:
          type: array
          items:
            type: object
            properties:
              salesPersonName:
                type: string
              leadsAssigned:
                type: integer
              ordersConverted:
                type: integer
              revenue:
                type: number
            required: [salesPersonName, leadsAssigned, ordersConverted, revenue]
        conversionRate:
          type: number
      required: [totalOrders, totalRevenue, bySalesPerson, conversionRate]
    FinanceReport:
      type: object
      properties:
        totalCollected:
          type: number
        totalPending:
          type: number
        totalOverdue:
          type: number
        daily:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              collected:
                type: number
            required: [date, collected]
        byPaymentMode:
          type: array
          items:
            type: object
            properties:
              mode:
                type: string
              amount:
                type: number
            required: [mode, amount]
      required: [totalCollected, totalPending, totalOverdue, daily, byPaymentMode]
    ServiceReport:
      type: object
      properties:
        totalCalls:
          type: integer
        closedCalls:
          type: integer
        avgResolutionDays:
          type: number
        byPriority:
          type: array
          items:
            type: object
            properties:
              priority:
                type: string
              count:
                type: integer
            required: [priority, count]
        byEngineer:
          type: array
          items:
            type: object
            properties:
              engineerName:
                type: string
              assigned:
                type: integer
              closed:
                type: integer
            required: [engineerName, assigned, closed]
      required: [totalCalls, closedCalls, avgResolutionDays, byPriority, byEngineer]
...
    Setting:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
      required: [key, value, updatedAt]
    SettingInput:
      type: object
      properties:
        value:
          type: string
      required: [value]
    BulkSettingsInput:
      type: object
      properties:
        settings:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              value:
                type: string
            required: [key, value]
      required: [settings]
...