# lib/api-zod/src/generated/types/leadsReport.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Solar Installation CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadsReportBySourceItem } from './leadsReportBySourceItem';
import type { LeadsReportByStageItem } from './leadsReportByStageItem';
import type { LeadsReportDailyItem } from './leadsReportDailyItem';

export interface LeadsReport {
  totalLeads: number;
  newLeads: number;
  convertedLeads: number;
  bySource: LeadsReportBySourceItem[];
  byStage: LeadsReportByStageItem[];
  daily: LeadsReportDailyItem[];
}